% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{assign_stat}
\alias{assign_stat}
\title{Assign stat}
\usage{
assign_stat(ma_df, weight_stat, seq_prop_stop = NULL, seq_alpha_stop = NULL)
}
\arguments{
\item{ma_df}{dataframe with the studies}

\item{weight_stat}{a vector with the expected weight for t_test, seq_t_test and bayes for each test in the metaanalysis}

\item{seq_prop_stop}{a vector with the proportions at which we will peek in the case of a sequential t test (Single value)}

\item{seq_prop_alpha}{a vector with the critical alpha value at each peek in the case of a sequential t test (Single Value)

For non-sequential t-test seq* parameters are set to NA}
}
\description{
Assign stat
}
