% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_power}
\alias{compute_power}
\title{Compute stats for posterior statistical tests}
\usage{
compute_power(m1, m2, n1, n2, v1, v2, df)
}
\arguments{
\item{m1}{mean of the group 1}

\item{m2}{mean of the group 2}

\item{n1}{number of participants in group 1}

\item{n2}{number of participants in group 2}

\item{v1}{variance of group 1}

\item{v2}{variance of group 2}

\item{df}{degrees of freedom}
}
\description{
Compute stats for posterior statistical tests
}
