% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_seq_t_test}
\alias{compute_seq_t_test}
\title{Compute a sequential t-test with a pre specified number of stops and a given alpha}
\usage{
compute_seq_t_test(study, prop_stop, alpha_stop)
}
\arguments{
\item{study}{a vector representing a study with 2n entries, n for each of the groups}

\item{prop_stop}{at which proportion of all data will we peek}

\item{alpha_stop}{alpha in each of the peeks}
}
\description{
Compute a sequential t-test with a pre specified number of stops and a given alpha
}
