% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_t_test}
\alias{compute_t_test}
\title{Compute a two-sample t-test}
\usage{
compute_t_test(study, tails = 1)
}
\arguments{
\item{study}{a study with one of the samples}

\item{tails}{number of tails}

\item{alpha}{alpha of the test}
}
\description{
We don't use the one by R to speed up the computation, it only computes the alternative of the difference being greater than 0 (i.e. one-tailed)
}
