% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ma.R
\name{genMA}
\alias{genMA}
\title{Generate a single group of metanalyses with the maximum number of participants}
\usage{
genMA(k, delta, tau, fixed.n = NULL)
}
\arguments{
\item{k}{number of studies in the metanalysis}

\item{delta}{true effect of the studies}

\item{tau}{the SD around the true effect}

\item{fixed.n}{number of participants *PER GROUP* (groups are always balanced), if null it is extracted from a distribution (see getN)}
}
\description{
Generate a single group of metanalyses with the maximum number of participants
}
