% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ma.R
\name{getN}
\alias{getN}
\title{Simulate a sample size}
\usage{
getN(k = 1, min.n = 5, max.n = 1905, shape = 1.15326986, scale = 0.04622745)
}
\description{
Legacy Meta-Showdown
}
\details{
get a simulated per-group sample size that follows the distribution of empirical sample sizes
}
