% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list2str}
\alias{list2str}
\title{Collapse a one level list into a string}
\usage{
list2str(l, sep = ":", collapse = " ")
}
\arguments{
\item{l}{list to be collapsed}

\item{sep}{same as in paste}

\item{collapse}{same as in paste}
}
\description{
Collapse a one level list into a string
}
