% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{run_seq_t_test}
\alias{run_seq_t_test}
\title{Run the seq test for all the studies in the dataframe}
\usage{
run_seq_t_test(ma_df)
}
\arguments{
\item{ma_df}{dataframe containing the metanalysis and the prop_stop and alpha_stop per study}
}
\description{
Run the seq test for all the studies in the dataframe
}
