% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{run_simulation}
\alias{run_simulation}
\title{Main simulation call}
\usage{
run_simulation(param_df, production = T)
}
\arguments{
\item{param_df, }{dataframe containing the parameters to call simulate_oneMA}

\item{production}{a boolean flag to disthinguish when we are in production or testing mode}
}
\description{
Main simulation call
}
