% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{run_stats}
\alias{run_stats}
\title{Run stats on each study}
\usage{
run_stats(ma_df_stat)
}
\arguments{
\item{ma_df_stat}{list of studies in the metanalysis with their stat assigned}
}
\description{
Run stats on each study
}
