% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{run_t_test}
\alias{run_t_test}
\title{Run the t-test for the studies in the dataframe}
\usage{
run_t_test(ma_df)
}
\arguments{
\item{ma_df}{dataframe containing the metanalysis}
}
\description{
Run the t-test for the studies in the dataframe
}
