% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_paramdf}
\alias{create_paramdf}
\title{Create a parameter dataframe}
\usage{
create_paramdf(k, delta, tau, fixed.n)
}
\arguments{
\item{k}{the number of studies in the MA (NULL if you w)}

\item{delta}{the true effect (or the average of the true effects if heterogeneity exists)}

\item{tau}{the SD around the true effect}
}
\description{
Create a parameter dataframe
}
