% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{run_stats}
\alias{run_stats}
\title{Run stats on each study}
\usage{
run_stats(ma_df, stat_type = c("t_test"))
}
\arguments{
\item{ma_df}{ma_df over which we want to apply the stats}

\item{stat_type}{type of stats we want to apply c=('t_test')}
}
\description{
Run stats on each study
}
